Extracting parsing out to its own thing
